Move write_line function to utils
